src/login-service/requirements.txt
flask
flask-login
pymongo
python-dotenv
dnspython